Visuals are set up

+++ FinalGame_Group5/Assets/keenan/scripts_keenan/Inventory_Final_Scripts/MainWeightController.cs
@@ -6,8 +6,8 @@ namespace keenan.scripts_keenan.Inventory_Final_Scripts
 {
     public class MainWeightController : MonoBehaviour
     {
-        public int maxWeight;
-        public int currWeight;
+        public float maxWeight;
+        public float currWeight;
 
         public Text weightTxt;
